requirements.txt
Django==5.0
django-environ==0.11.2
psycopg2==2.9.9
djangorestframework==3.14.0